+++ restli-client/src/main/java/com/linkedin/restli/client/Request.java
@@ -19,6 +19,7 @@ package com.linkedin.restli.client;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
+import com.linkedin.data.DataMap;
 import com.linkedin.data.schema.PathSpec;
 import com.linkedin.data.template.RecordTemplate;
 import com.linkedin.data.transform.filter.request.MaskTree;
@@ -323,6 +324,11 @@ public class Request<T>
       {
         throw new IllegalArgumentException("Field param was a string and it did not represent a serialized mask tree", e);
       }
+    }
+    else if (fields instanceof DataMap)
+    {
+      MaskTree tree = new MaskTree((DataMap) fields);
+      return tree.getOperations().keySet();
     }
 
     throw new IllegalArgumentException("Fields param is of unrecognized type: " + fields.getClass());