+++ CHANGELOG.md
@@ -13,7 +13,17 @@ When updating the changelog, remember to be very clear about what behavior has c
 and what APIs have changed, if applicable.
 
 ## [Unreleased]
+
+## [29.17.0] - 2021-03-23
+- Implement D2 cluster subsetting.
+
+## [29.16.2] - 2021-03-22
+- Fix an issue where in collection response, we did not fill in the default values in the metadata and paging metadata.
+
+## [29.16.1] - 2021-03-17
 - Add fluent client api for simple resource and association resource.
+- Add support for generating projection mask as the mask data map.
+- Fix UnmodifiableList wrap in d2 relative load balancer
 
 ## [29.16.0] - 2021-03-10
 - Add a ParSeq based CompletionStage implementation
@@ -4882,7 +4892,10 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.16.0...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.17.0...master
+[29.17.0]: https://github.com/linkedin/rest.li/compare/v29.16.2...v29.17.0
+[29.16.2]: https://github.com/linkedin/rest.li/compare/v29.16.1...v29.16.2
+[29.16.1]: https://github.com/linkedin/rest.li/compare/v29.16.0...v29.16.1
 [29.16.0]: https://github.com/linkedin/rest.li/compare/v29.15.9...v29.16.0
 [29.15.9]: https://github.com/linkedin/rest.li/compare/v29.15.8...v29.15.9
 [29.15.8]: https://github.com/linkedin/rest.li/compare/v29.15.7...v29.15.8